+++ src/models/serialization_status.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class SerializationStatus(Enum):
+    ONGOING = "연재"
+    COMPLETED = "완결"
+    HIATUS = "휴재"
